requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
stripe==7.8.0
twilio==8.10.0
sendgrid==6.10.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
dj-database-url==2.1.0